Create Queue
Reasons for using linked lists to implement queues:

Dynamic size: The queue can grow and shrink dynamically, unlike with arrays.
No shifting: The front element of the queue can be removed (enqueue) without having to shift other elements in the memory.
Reasons for not using linked lists to implement queues:

Extra memory: Each queue element must contain the address to the next element (the next linked list node).
Readability: The code might be harder to read and write for some because it is longer and more complex.